+++ src/Data/Morpheus/Types/Internal/Resolving/Resolver.hs
@@ -21,7 +21,7 @@ module Data.Morpheus.Types.Internal.Resolving.Resolver
   , Resolver(..)
   , ResolvingStrategy(..)
   , MapStrategy(..)
-  , LiftEither(..)
+  , LiftOperation
   , resolveObject
   , resolveEnum
   , toResponseRes
@@ -96,11 +96,11 @@ import           Data.Morpheus.Types.IO         ( renderResponse
                                                 )
 -- MORPHEUS
 
-class LiftEither (o::OperationType) res where
+class LiftOperation (o::OperationType) res where
   type ResError res :: *
-  liftEither :: Monad m => m (Either (ResError res) a) -> res o event m  a
+  liftOperation :: Monad m => m (Either (ResError res) a) -> res o event m  a
 
-type WithOperation (o :: OperationType) = LiftEither o Resolver
+type WithOperation (o :: OperationType) = LiftOperation o Resolver
 
 type ResponseStream event m = ResultT (ResponseEvent m event) GQLError 'True m
 
@@ -142,8 +142,8 @@ instance Monad m => Functor (ResolvingStrategy o e m) where
   fmap f (ResolveS res) = ResolveS $ (<$>) f <$> res
 
 -- Applicative
-instance (LiftEither o ResolvingStrategy, Monad m) => Applicative (ResolvingStrategy o e m) where
-  pure = liftEither . pure . pure
+instance (LiftOperation o ResolvingStrategy, Monad m) => Applicative (ResolvingStrategy o e m) where
+  pure = liftOperation . pure . pure
   -------------------------------------
   (ResolveQ f) <*> (ResolveQ res) = ResolveQ (f <*> res)
   ------------------------------------------------------------------------
@@ -151,22 +151,22 @@ instance (LiftEither o ResolvingStrategy, Monad m) => Applicative (ResolvingStra
   --------------------------------------------------------------
   (ResolveS f) <*> (ResolveS res) = ResolveS $ (<*>) <$> f <*> res
 
--- LiftEither
-instance LiftEither QUERY ResolvingStrategy where
+-- LiftOperation
+instance LiftOperation QUERY ResolvingStrategy where
   type ResError ResolvingStrategy = GQLErrors
-  liftEither = ResolveQ . ResultT . fmap fromEither
+  liftOperation = ResolveQ . ResultT . fmap fromEither
 
-instance LiftEither MUTATION ResolvingStrategy where
+instance LiftOperation MUTATION ResolvingStrategy where
   type ResError ResolvingStrategy = GQLErrors
-  liftEither = ResolveM . ResultT . fmap fromEither
+  liftOperation = ResolveM . ResultT . fmap fromEither
 
-instance LiftEither SUBSCRIPTION ResolvingStrategy where
+instance LiftOperation SUBSCRIPTION ResolvingStrategy where
   type ResError ResolvingStrategy = GQLErrors
-  liftEither = ResolveS . ResultT . fmap (fromEither . fmap pure)
+  liftOperation = ResolveS . ResultT . fmap (fromEither . fmap pure)
 
  -- Failure
-instance (LiftEither o ResolvingStrategy, Monad m) => Failure GQLErrors (ResolvingStrategy o e m) where
-  failure = liftEither . pure . Left
+instance (LiftOperation o ResolvingStrategy, Monad m) => Failure GQLErrors (ResolvingStrategy o e m) where
+  failure = liftOperation . pure . Left
 
 -- DataResolver
 data DataResolver o e m =
@@ -181,7 +181,7 @@ instance Semigroup (DataResolver o e m) where
   _           <> _           = InvalidRes "can't merge: incompatible resolvers"
 
 withObject
-  :: (LiftEither o ResolvingStrategy, Monad m)
+  :: (LiftOperation o ResolvingStrategy, Monad m)
   => (ValidSelectionSet -> ResolvingStrategy o e m value)
   -> (Key, ValidSelection)
   -> ResolvingStrategy o e m value
@@ -191,7 +191,7 @@ withObject _ (key, Selection { selectionPosition }) =
   failure (subfieldsNotSelected key "" selectionPosition)
 
 resolveObject
-  :: (Monad m, LiftEither o ResolvingStrategy)
+  :: (Monad m, LiftOperation o ResolvingStrategy)
   => ValidSelectionSet
   -> DataResolver o e m
   -> ResolvingStrategy o e m ValidValue
@@ -207,7 +207,7 @@ resolveObject _ _ =
   failure $ internalResolvingError "expected object as resolver"
 
 resolveEnum
-  :: (Monad m, LiftEither o ResolvingStrategy)
+  :: (Monad m, LiftOperation o ResolvingStrategy)
   => Name
   -> Name
   -> ValidSelectionRec
@@ -228,7 +228,7 @@ resolveEnum _ _ _ =
 
 
 resolve__typename
-  :: (Monad m, LiftEither o ResolvingStrategy)
+  :: (Monad m, LiftOperation o ResolvingStrategy)
   => Name
   -> (Key, (Key, ValidSelection) -> ResolvingStrategy o e m ValidValue)
 resolve__typename name = ("__typename", const $ pure $ gqlString name)
@@ -276,8 +276,8 @@ instance Functor m => Functor (Resolver o e m) where
     where eventFmap res event = fmap f (res event)
 
 -- Applicative
-instance (LiftEither o Resolver ,Monad m) => Applicative (Resolver o e m) where
-  pure = liftEither . pure . pure
+instance (LiftOperation o Resolver ,Monad m) => Applicative (Resolver o e m) where
+  pure = liftOperation . pure . pure
   -------------------------------------
   (QueryResolver f) <*> (QueryResolver res) = QueryResolver (f <*> res)
   ---------------------------------------------------------------------
@@ -303,27 +303,27 @@ instance (Monad m) => Monad (Resolver MUTATION e m) where
 
 -- Monad Transformers    
 instance MonadTrans (Resolver QUERY e) where
-  lift = liftEither . fmap pure
+  lift = liftOperation . fmap pure
 
 instance MonadTrans (Resolver MUTATION e) where
-  lift = liftEither . fmap pure
+  lift = liftOperation . fmap pure
 
--- LiftEither
-instance LiftEither QUERY Resolver where
+-- LiftOperation
+instance LiftOperation QUERY Resolver where
   type ResError Resolver = String
-  liftEither = QueryResolver . ResultT . fmap fromEitherSingle
+  liftOperation = QueryResolver . ResultT . fmap fromEitherSingle
 
-instance LiftEither MUTATION Resolver where
+instance LiftOperation MUTATION Resolver where
   type ResError Resolver = String
-  liftEither = MutResolver . ResultT . fmap (fromEitherSingle . fmap ([], ))
+  liftOperation = MutResolver . ResultT . fmap (fromEitherSingle . fmap ([], ))
 
-instance LiftEither SUBSCRIPTION Resolver where
+instance LiftOperation SUBSCRIPTION Resolver where
   type ResError Resolver = String
-  liftEither = SubResolver [] . const . liftEither
+  liftOperation = SubResolver [] . const . liftOperation
 
 -- Failure
-instance (LiftEither o Resolver, Monad m) => Failure Message (Resolver o e m) where
-  failure = liftEither . pure . Left . unpack
+instance (LiftOperation o Resolver, Monad m) => Failure Message (Resolver o e m) where
+  failure = liftOperation . pure . Left . unpack
 
 -- Type Helpers  
 type family UnSubResolver (a :: * -> *) :: (* -> *)
@@ -336,7 +336,7 @@ type FieldRes o e m
   = (Key, (Key, ValidSelection) -> ResolvingStrategy o e m ValidValue)
 
 toResolver
-  :: (LiftEither o Resolver, Monad m)
+  :: (LiftOperation o Resolver, Monad m)
   => Validation a
   -> (a -> Resolver o e m b)
   -> Resolver o e m b